+++ cardano-crypto-class/src/Cardano/Crypto/KES/Class.hs
@@ -94,7 +94,11 @@ import NoThunks.Class (NoThunks)
 
 import Cardano.Binary (Decoder, Encoding, Size, decodeBytes, encodeBytes, withWordSize)
 
-import Cardano.Crypto.DSIGN.Class (failSizeCheck)
+import Cardano.Binary.FixedSizeCodec (
+  FixedSizeCodec (..),
+  decodeFixedSized,
+  encodeFixedSized,
+ )
 import Cardano.Crypto.Hash.Class (Hash, HashAlgorithm, hashWith)
 import Cardano.Crypto.Libsodium (MLockedAllocator, mlockedMalloc)
 import Cardano.Crypto.Libsodium.MLockedSeed
@@ -119,6 +123,8 @@ class
   , KnownNat (SignKeySizeKES v)
   , KnownNat (SigSizeKES v)
   , KnownNat (TotalPeriodsKES v)
+  , FixedSizeCodec (VerKeyKES v)
+  , FixedSizeCodec (SigKES v)
   ) =>
   KESAlgorithm v
   where
@@ -131,8 +137,11 @@ class
 
   type SeedSizeKES v :: Nat
   type VerKeySizeKES v :: Nat
+  type VerKeySizeKES v = FixedSize (VerKeyKES v)
   type SignKeySizeKES v :: Nat
+  type SignKeySizeKES v = FixedSize (SignKeyKES v)
   type SigSizeKES v :: Nat
+  type SigSizeKES v = FixedSize (SigKES v)
   type TotalPeriodsKES v :: Nat
 
   type SizeVerKeyKES v :: Nat
@@ -194,10 +203,14 @@ class
   --
 
   rawSerialiseVerKeyKES :: VerKeyKES v -> ByteString
+  rawSerialiseVerKeyKES = rawEncodeFixedSized
   rawSerialiseSigKES :: SigKES v -> ByteString
+  rawSerialiseSigKES = rawEncodeFixedSized
 
   rawDeserialiseVerKeyKES :: ByteString -> Maybe (VerKeyKES v)
+  rawDeserialiseVerKeyKES = rawDecodeFixedSized
   rawDeserialiseSigKES :: ByteString -> Maybe (SigKES v)
+  rawDeserialiseSigKES = rawDecodeFixedSized
 
   deriveVerKeyKES :: (MonadST m, MonadThrow m) => SignKeyKES v -> m (VerKeyKES v)
 
@@ -323,6 +336,7 @@ updateKES = updateKESWith mlockedMalloc
 class
   ( KESAlgorithm v
   , NoThunks (UnsoundPureSignKeyKES v)
+  , FixedSizeCodec (UnsoundPureSignKeyKES v)
   ) =>
   UnsoundPureKESAlgorithm v
   where
@@ -359,7 +373,9 @@ class
     m (SignKeyKES v)
 
   rawSerialiseUnsoundPureSignKeyKES :: UnsoundPureSignKeyKES v -> ByteString
+  rawSerialiseUnsoundPureSignKeyKES = rawEncodeFixedSized
   rawDeserialiseUnsoundPureSignKeyKES :: ByteString -> Maybe (UnsoundPureSignKeyKES v)
+  rawDeserialiseUnsoundPureSignKeyKES = rawDecodeFixedSized
 
 -- | Unsound operations on KES sign keys. These operations violate secure
 -- forgetting constraints by leaking secrets to unprotected memory. Consider
@@ -463,12 +479,15 @@ instance
 
 encodeVerKeyKES :: KESAlgorithm v => VerKeyKES v -> Encoding
 encodeVerKeyKES = encodeBytes . rawSerialiseVerKeyKES
+{-# DEPRECATED encodeVerKeyKES "Use `encodeFixedSized` instead" #-}
 
 encodeUnsoundPureSignKeyKES :: UnsoundPureKESAlgorithm v => UnsoundPureSignKeyKES v -> Encoding
 encodeUnsoundPureSignKeyKES = encodeBytes . rawSerialiseUnsoundPureSignKeyKES
+{-# DEPRECATED encodeUnsoundPureSignKeyKES "Use `encodeFixedSized` instead" #-}
 
 encodeSigKES :: KESAlgorithm v => SigKES v -> Encoding
 encodeSigKES = encodeBytes . rawSerialiseSigKES
+{-# DEPRECATED encodeSigKES "Use `encodeFixedSized` instead" #-}
 
 encodeSignKeyKES ::
   forall v m.
@@ -478,29 +497,20 @@ encodeSignKeyKES ::
 encodeSignKeyKES = fmap encodeBytes . rawSerialiseSignKeyKES
 
 decodeVerKeyKES :: forall v s. KESAlgorithm v => Decoder s (VerKeyKES v)
-decodeVerKeyKES = do
-  bs <- decodeBytes
-  case rawDeserialiseVerKeyKES bs of
-    Just vk -> return vk
-    Nothing -> failSizeCheck "decodeVerKeyKES" "key" bs (verKeySizeKES (Proxy :: Proxy v))
+decodeVerKeyKES = decodeFixedSized
 {-# INLINE decodeVerKeyKES #-}
+{-# DEPRECATED decodeVerKeyKES "Use `decodeFixedSized` instead" #-}
 
 decodeUnsoundPureSignKeyKES ::
   forall v s. UnsoundPureKESAlgorithm v => Decoder s (UnsoundPureSignKeyKES v)
-decodeUnsoundPureSignKeyKES = do
-  bs <- decodeBytes
-  case rawDeserialiseUnsoundPureSignKeyKES bs of
-    Just vk -> return vk
-    Nothing -> failSizeCheck "decodeUnsoundPureSignKeyKES" "key" bs (signKeySizeKES (Proxy :: Proxy v))
+decodeUnsoundPureSignKeyKES = decodeFixedSized
 {-# INLINE decodeUnsoundPureSignKeyKES #-}
+{-# DEPRECATED decodeUnsoundPureSignKeyKES "Use `decodeFixedSized` instead" #-}
 
 decodeSigKES :: forall v s. KESAlgorithm v => Decoder s (SigKES v)
-decodeSigKES = do
-  bs <- decodeBytes
-  case rawDeserialiseSigKES bs of
-    Just sig -> return sig
-    Nothing -> failSizeCheck "decodeSigKES" "signature" bs (sigSizeKES (Proxy :: Proxy v))
+decodeSigKES = decodeFixedSized
 {-# INLINE decodeSigKES #-}
+{-# DEPRECATED decodeSigKES "Use `decodeFixedSized` instead" #-}
 
 decodeSignKeyKES ::
   forall v s m.
@@ -534,6 +544,12 @@ deriving instance KESAlgorithm v => Eq (SignedKES v a)
 
 instance KESAlgorithm v => NoThunks (SignedKES v a)
 
+instance KESAlgorithm v => FixedSizeCodec (SignedKES v a) where
+  type FixedSize (SignedKES v a) = FixedSize (SigKES v)
+
+  rawEncodeFixedSized (SignedKES x) = rawEncodeFixedSized x
+  rawDecodeFixedSized bs = SignedKES <$> rawDecodeFixedSized bs
+
 -- use generic instance
 
 signedKES ::
@@ -565,11 +581,13 @@ unsoundPureSignedKES ::
 unsoundPureSignedKES ctxt time a key = SignedKES $ unsoundPureSignKES ctxt time a key
 
 encodeSignedKES :: KESAlgorithm v => SignedKES v a -> Encoding
-encodeSignedKES (SignedKES s) = encodeSigKES s
+encodeSignedKES = encodeFixedSized
+{-# DEPRECATED encodeSignedKES "Use `encodeFixedSized` instead" #-}
 
 decodeSignedKES :: KESAlgorithm v => Decoder s (SignedKES v a)
-decodeSignedKES = SignedKES <$> decodeSigKES
+decodeSignedKES = decodeFixedSized
 {-# INLINE decodeSignedKES #-}
+{-# DEPRECATED decodeSignedKES "Use `decodeFixedSized` instead" #-}
 
 -- | A sign key bundled with its associated period.
 data SignKeyWithPeriodKES v = SignKeyWithPeriodKES