+++ cardano-crypto-class/src/Cardano/Crypto/PinnedSizedBytes.hs
@@ -1,6 +1,8 @@
+{-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DerivingVia #-}
+{-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE KindSignatures #-}
@@ -9,6 +11,7 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TemplateHaskellQuotes #-}
 {-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UnboxedTuples #-}
 
 module Cardano.Crypto.PinnedSizedBytes (
@@ -23,6 +26,7 @@ module Cardano.Crypto.PinnedSizedBytes (
   psbToByteArray,
   psbFromByteString,
   psbFromByteStringM,
+  psbFromByteStringForM,
   psbFromByteStringCheck,
   psbToByteString,
 
@@ -81,11 +85,20 @@ import qualified Data.ByteString as BS
 import qualified Data.ByteString.Short as SBS
 import qualified Data.Primitive as Prim
 
+import Cardano.Binary.FixedSizeCodec (
+  FixedSizeCodec (..),
+  decodeFixedSized,
+  encodeFixedSized,
+  guardFixedSized,
+ )
 import Cardano.Crypto.Libsodium.C (c_sodium_compare)
 import Cardano.Crypto.PackedBytes.Internal (PackedBytes, packBytes)
 import Cardano.Crypto.Util (decodeHexString)
 import Cardano.Foreign
+import Control.Monad.Trans.Fail (errorFail)
+import qualified Data.ByteString.Unsafe as BS
 import Data.MemPack.Buffer (byteArrayToShortByteString)
+import GHC.Stack (HasCallStack)
 
 {- HLINT ignore "Reduce duplication" -}
 
@@ -135,7 +148,7 @@ instance KnownNat n => Ord (PinnedSizedBytes n) where
       size = fromInteger (natVal (Proxy :: Proxy n))
 
 instance KnownNat n => ToCBOR (PinnedSizedBytes n) where
-  toCBOR = toCBOR . psbToByteString
+  toCBOR = encodeFixedSized
   {-# INLINE toCBOR #-}
 
   encodedSizeExpr _size proxy =
@@ -145,17 +158,13 @@ instance KnownNat n => ToCBOR (PinnedSizedBytes n) where
       pinnedBytesSize = fromInteger (natVal (Proxy @n))
 
 instance KnownNat n => FromCBOR (PinnedSizedBytes n) where
-  fromCBOR = do
-    bs <- fromCBOR
-    case psbFromByteStringCheck bs of
-      Nothing ->
-        fail $
-          "Size mismatch. Expected: " <> show size <> " number of bytes, but got: " <> show (BS.length bs)
-      Just psb -> pure psb
-    where
-      size :: Int
-      size = fromInteger (natVal (Proxy :: Proxy n))
+  fromCBOR = decodeFixedSized
   {-# INLINE fromCBOR #-}
+
+instance KnownNat n => FixedSizeCodec (PinnedSizedBytes n) where
+  type FixedSize (PinnedSizedBytes n) = n
+  rawEncodeFixedSized = psbToByteString
+  rawDecodeFixedSized = psbFromByteStringM
 
 -- | This instance is meant to be used with @TemplateHaskell@
 --
@@ -180,7 +189,14 @@ instance KnownNat n => IsString (Q (TExp (PinnedSizedBytes n))) where
     let n = fromInteger $ natVal (Proxy :: Proxy n)
     case decodeHexString hexStr n of
       Left err -> fail $ "<PinnedSizedBytes>: " ++ err
-      Right _ -> examineSplice [||either error psbFromByteString (decodeHexString hexStr n)||]
+      Right _ ->
+        examineSplice
+          [||
+          either
+            error
+            (errorFail @String . rawDecodeFixedSized)
+            (decodeHexString hexStr n)
+          ||]
 
 instance KnownNat n => IsString (Code Q (PinnedSizedBytes n)) where
   fromString = Code . fromString
@@ -226,37 +242,31 @@ psbFromBytes ws0 = PSB (pinnedByteArrayFromListN size ws)
 
 -- | Convert a ByteString into PinnedSizedBytes. Input should contain the exact
 -- number of bytes as expected by type level @n@ size, otherwise error.
-psbFromByteString :: KnownNat n => BS.ByteString -> PinnedSizedBytes n
-psbFromByteString bs =
-  case psbFromByteStringCheck bs of
-    Nothing -> error $ "psbFromByteString: Size mismatch, got: " ++ show (BS.length bs)
-    Just psb -> psb
+psbFromByteString :: (HasCallStack, KnownNat n) => BS.ByteString -> PinnedSizedBytes n
+psbFromByteString = errorFail @String . psbFromByteStringM
 
-psbFromByteStringCheck :: forall n. KnownNat n => BS.ByteString -> Maybe (PinnedSizedBytes n)
+psbFromByteStringForM ::
+  forall a m.
+  ( MonadFail m
+  , KnownNat (FixedSize a)
+  ) =>
+  Proxy a -> BS.ByteString -> m (PinnedSizedBytes (FixedSize a))
+psbFromByteStringForM _ = psbFromByteStringM @(FixedSize a)
+
+psbFromByteStringCheck :: KnownNat n => BS.ByteString -> Maybe (PinnedSizedBytes n)
 psbFromByteStringCheck = psbFromByteStringM
 
 psbFromByteStringM ::
   forall n m.
-  (KnownNat n, MonadFail m) =>
+  (MonadFail m, KnownNat n) =>
   BS.ByteString -> m (PinnedSizedBytes n)
-psbFromByteStringM bs
-  | n == size = pure $
-      unsafeDupablePerformIO $
-        BS.useAsCStringLen bs $ \(Ptr addr#, _) -> do
-          marr@(MutableByteArray marr#) <- newPinnedByteArray size
-          primitive_ $ copyAddrToByteArray# addr# marr# 0# (case size of I# s -> s)
-          arr <- unsafeFreezeByteArray marr
-          return (PSB arr)
-  | otherwise =
-      fail $
-        "Supplied ByteString with size: "
-          <> show n
-          <> " did not match the expected number of bytes: "
-          <> show size
-  where
-    n = BS.length bs
-    size :: Int
-    size = fromInteger (natVal (Proxy :: Proxy n))
+psbFromByteStringM bs =
+  guardFixedSized bs $ do
+    unsafeDupablePerformIO . BS.unsafeUseAsCStringLen bs $ \(Ptr addr#, size) -> do
+      marr@(MutableByteArray marr#) <- newPinnedByteArray size
+      primitive_ $ copyAddrToByteArray# addr# marr# 0# (case size of I# s -> s)
+      arr <- unsafeFreezeByteArray marr
+      return (pure $ PSB arr)
 
 {-# DEPRECATED psbZero "This is not referentially transparent" #-}
 psbZero :: KnownNat n => PinnedSizedBytes n